pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dubito"
version = "0.0.1"
description = "Subito tracker is a tool to track the price of a product on subito.it"
readme = "README.md"
authors = [{ name = "Valerio Di Pasquale", email = "valeriodipasq@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["subito", "scraper", "dubito", "scraping"]
dependencies = [
    "selectorlib == 0.16.0",
    "requests",
    "requests-cache == 1.0.1",
    "pandas == 2.0.1",
    "validators == 0.20.0",
    "rich == 13.6.0",
    "seaborn == 0.13.0",
    "matplotlib == 3.8.0",
    "mplfinance == 0.12.10b0",
    "ta == 0.11.0",
    "newspaper3k == 0.2.8",
    "wordcloud == 1.9.3", # TODO: remove in future versions
    "deep-translator == 1.11.4", # TODO: remove in future versions
    "nltk == 3.8.1",
    "peewee==3.17.0"
]
requires-python = ">=3.7"

[project.urls]
Home = "https://github.com/daqh/dubito"

[project.scripts]
dubito = "dubito.__main__:main"
